Add Django 1.8 and 1.9 test projects with apps
Although these projects can be created programmatically with make, I've
committed them for stability. I'm imagining a test bug which could be
really hard to debug if it arose because of how a particular computer
was invoking the make recipes, or something wasn't generated in exactly
the same way as it is here.

Therefore, current strategy is: is make to build the projects, but
commit them to the repo for stability / debugging.